fix: address code review findings for JWT/forward-auth feature
- Forward-auth middleware now skips if X-JWT header is already set by
  JWT middleware, preventing silent identity override when both are enabled
- Strip forward-auth header after consuming it to avoid leaking user
  claims JSON downstream through the handler chain
- Add startup validation that AUTH_JWT_HEADER is non-empty when
  AUTH_JWT_SECRET is set
- Fix TestClaimStr to use float64(42) matching actual JSON unmarshal
  behavior instead of Go int
- Document that external JWT tokens must include exp (expiration) claim

+++ backend/app/rest/api/jwt_auth_test.go
@@ -714,7 +714,7 @@ func TestExtractUser(t *testing.T) {
 func TestClaimStr(t *testing.T) {
 	claims := map[string]interface{}{
 		"string_val": "hello",
-		"int_val":    42,
+		"int_val":    float64(42), // JSON always unmarshals numbers as float64
 		"float_val":  3.14,
 		"bool_val":   true,
 	}
